v4/go: apply the slice-N round-1 review fixes (SOW-0025 chunk 4-8)
Rust parity: the snapshot identity compare moves before any
construction (the attempt file is still empty when the refusal
fires, matching api.rs), and the live self-replacement probe
classifies a symlinked destination with the Rust-verbatim symlink
detail. The resolution mode becomes a closed enum: an out-of-range
mode refuses with the invalid-argument class before any path access
(Rust exhaustive match).

Go idioms: the PublishAttempt enforces its consume rule like the
residue handle (file exposure goes nil on every terminal, a second
Finish refuses), the dead public-to-internal coordination converter
is removed, the reservation policy/phase evidence maps through
explicit switches instead of numeric casts, the stale
HousekeepingArtifact comment restates the design, and the root test
uses errors.As with the correct temp-name capacity.

+++ v4/go/internal/mapping/mapping_identity_probe.go
@@ -18,5 +18,5 @@ func StatIdentity(path string) (device uint64, inode uint64, err error) {
 	if err := unix.Stat(path, &st); err != nil {
 		return 0, 0, &format.Error{Code: format.CodeIO, Detail: "publication filesystem operation failed"}
 	}
-	return st.Dev, st.Ino, nil
+	return uint64(st.Dev), uint64(st.Ino), nil
 }
